Use separate identities for storage and networking.
Because it turns out that network identities can be quite fickle in Java, decouple any network-based identity from the storage identity entirely. This will continue to use the UUID from JMRI 4.14 as the storage identity if available, otherwise it will just use a random UUID. If a reliable cross platform mechanism for ensuring a machine's identity is available without relying on the contents of a user's home directory or network interfaces becomes available, it would be preferable to change to using that; therefore store the JMRI 4.14 compatible UUID as a separate element from the storage identity in use.