Fixing issues with parsing of turnout response from DCC++.
The default parsing was incorrectly identifying the turnout response
from DCC++ as THROWN when in fact the response was CLOSED. This was due
to the order in which the regex was checked and the incorrect index for
the regex in the getTOStateInt method. Sample log before / after below:

BEFORE:
[RX: H 1 1 0 0]   Turnout Reply:
	T/O Number: 1
	Direction: THROWN
[RX: H 2 2 1 1]   Turnout Reply:
	T/O Number: 2
	Direction: CLOSED

AFTER:
[RX: H 1 1 0 0]   Turnout Reply:
	T/O Number: 1
	T/O Address: 1
	T/O Index: 0
	T/O DCC Address: 1
	Direction: CLOSED
[RX: H 2 2 1 0]   Turnout Reply:
	T/O Number: 2
	T/O Address: 2
	T/O Index: 1
	T/O DCC Address: 6
	Direction: CLOSED

I've added the DCC Address as calculated from the address / index pair
to the output as well. This may help people with knowing the addressing
scheme used by DCC++ (based on DCC packet format).

I've also corrected the pretty print for output command responses as it
was listing all entries as OutputState.