Be nice to users who upgrade from 4.14
Use existing JMRI 4.14 local profile data if it exists, but use newer style otherwise.

(Because I lazyly reused the network identity to provide per-machine uniqueness for profile preferences, and that form was reused when breaking the network identities, the form for identifying a local storage in a profile nonsensically contains the machine identity and profile identity in 4.14 and earlier. The new form contains only the machine identity.)