Fire changes in correct order when deleting a bean
When a ProxyManager is the default manager, any listeners to it *also* become listeners to the underlying Managers. When a bean is deleted, the Managers fire a propertyChange such that when an Object listening for those changes from the default manager requested a list of beans from the default manager, it would get a list that contained the removed bean, but if on a different thread, could have the list change while iterating over it.

This change ensures the ProxyManagers for a Manager rebuild their list of beans before the PropertyChangeEvents are sent (it also includes some automagic changes from Eclipse).