Add connection configurations for Digitrax DCS240, DCS52, PR4 (#6599)
Adds specific connection configurations for newer Digitrax USB hardware.  

Adds "decoder" definition for DCS52's OpSws.  Adds and updates a variety of help files.  

Fixes a couple of potential NPEs when a DCS52 or DCS240 is used in "Standalone programmer" mode and a "track power" control message is seen.

Changes how menu items are presented in the LocoNet menu - does not display items which aren't useful when configured for an interface in "standalone programer" mode.

Changes the order in which items are presented for the "Command station type":

- For an interface with "standalone programmer" mode capability, prioritize that mode to the top of the list;

- For an interface with built-in "command station" capability, prioritize the associated "command station" mode to be just below the associated "standalone programmer" mode.

Updates various help pages to reflect functionality of the PR4, DCS52, and DCS240.