chore: Silence ~201 constraint violation warnings
`Comparable.compareTo(Object)` has a specific contract for dealing with null arguments. By removing the nonNull annotation and explictly throwning an NPE `NamedBean.compareTo(Object)` adheres to the contract it is implementing, and a number of warnings in code about not adhering to the contract of an overridden method are silenced.