feat: Allow get parameters to be used
Allow all JSON services to use parameters (HTTP URL parameters in the RESTful server or as contents of a data body in sockets) to influence the data that is returned.

Use the `server=true|false` parameter in the schema server to specify if both server and client schemas are returned (no parameter) or only the server schema is returned (`true`) or only client schema is returned (`false`).