feat: Add message id to JSON Socket messages
JSON clients can add an id to messages to the JSON server that will be returned unchanged in all messages that are a direct response to the message sent. It is recommended clients using the id always increment the id after sending a message.

Breaking change: JSON Operations use of "id" property and JSON Throttle use of "throttle" property have both been replaced with use of "name" property. Use of "id" in JSON Throttle has been replaced with "rosterEntry". Where "name" was used in combination with "id", "username" is now used for old use of "name".

Breaking change: many JSON handling messages have had an id parameter added without deprecating the id-less version of the method, since any dropping of the id would violate the contract for use of the JSON id property.