fix: correct how listeners are tracked for JSON sockets
When using names to keep track of listeners, If a bean was renamed, the client would then request it, and the bean would get two listeners (one for each name) with double the needed traffic; in addition, if the name were then used on a new bean, the new bean would not be listened to since its name was already being listened to. The solution to these concerns is to use the bean, not its name, as the key of the map of listeners.