fix: throw exception on invalid string for color
In `ColorUtil.stringToColor()` throw an IllegalArgumentException if the String input cannot be turned into a color. Previously this method returned `Color.BLACK` when unable to parse a color string. Xml loaders all catch the exception and set the color being set at the time to black to avoid blocking errors loading files.