Change AnalogIO to use double
When I created the AnalogIO interface, I thought that float is enough. But I have later realized that it would be better if AnalogIO uses double instead of float.

This changes the public interface of AnalogIO, but since AnalogIO is added lately, it's probably none using it outside of JMRI yet.