Fixes bug in OlcbSensor that caused messages sent to the OpenLCB bus out of order.
When a sensor is changed, the sequence of enqueue-output-message and notify-listeners was
incorrect (my fault, sorry), which caused problems in the following scenario:

A sensor T that is being changed locally by JMRI (e.g. due to a panel icon action), which
triggers a Logix, and in that Logix there is an action that sets a second Sensor U.
The correct order of outgoing messages is T=[new state] U=[new state], i.e., a prefix
traversal of the listener tree. Due to the bug a postfix traversal of the notification
tree was happening, causing the triggering action T=[new state] to be output last to the bus.