fix: validate must not log, so remove logging switch
Because of the frequency with which validation methods get called with known invalid input (i.e. as a user is typing to get to a valid input), the validation methods must not log invalid inputs at a normal level (i.e.  it may eventually be deirable to have a validation routine log at the debug or trace level, but they should never log at the info, warning, or error levels).