fix: Exporting a profile no longer throws an NPE
Exporting a Profile no longer throws an NPE, although some older Profiles, never used with JMRI 4.12 or newer, may not export correctly.

This is involved because the way this was fixed was not to re-implement how Profiles are read, but to rely on the multi-profile design of the PreferencesManager to allow preferences from any Profile to be loaded. That decision led to fixing two PrefencesManager implementations to be multi-profile, which led to making FileUtil multi-profile capable.