fix: correctly compare NamedBeans when not all beans have user names
Fixes #7302 by only comparing the system names of two NamedBeans when neither NamedBean has a user name; otherwise this compares the user name of a NamedBean against the system name of a NamedBean if one of the NamedBeans does not have a user name. This does not use the display name of a NamedBean to avoid possibly attempting to sort two NamedBeans without user names incorrectly.