LnSensorManager deprecation of one sensor numbering format (#7377)
Deprecate LocoNetSensorManager's partial support for createSensorName("3:15","L") where "3:15" is supposed to mean "board 3, bit 15", and which returns the equivalent sensor system name of "LS47".  

Logs a one-time (per JMRI session) warning to the System Console log stating that the format is deprecated and support will be removed at a future date.