Refresh LocoNet slot data when loco reset button pushed (#7449)
Trigger a refresh of the slot information when the command station "Loco Reset" button is pushed and the feature is enabled.

At this time, this feature is implemented only by DCS240 and DCS210 command stations. JMRI will only refresh its slot information when one of those command stations is in-use, with Loco Reset functionality enabled, when the JMRI connection is configured to use one of those command stations.