test: Ignore tests failing consistently
See #7633. These tests fail when explicitly validating that a json-schema is a json-schema valid json-schema (other tests implicitely test this, so there are no worries about creating invalid json-schema).

The immediate cause of this is that the Java URLConnection class will not follow an HTTP 301 redirect from the http to the https redirect.

This has also revealed that when a json-schema defines an object in a JSON message to be a json-schema, the json schema validator ignores the mapping to not use the remote resource. Because this needs to be fixed upstream, tests that JSON messages containing json-schema are valid need to be ignored until this is fixed upstream.