feat: Require that Routes have valid System Names
This uses __O__ for the type letter for Routes since __R__ is also used by Reporters, and changing Routes is easier than changing Reporters.

When loading an XML file, Routes with names starting with connection prefix and the type letter _R_ get converted to using the type letter _O_, while Routes with names starting with no connection prefix get the connection prefix and type letter prepended.

The user gets informed that the changed occured with a simple dialog and warnings in the log.