test: Add property to silence dialogs in specific tests
This is a temporary fix to allow tests to continue past a modal dialog. A better (but too complex for this PR) fix would be to have a mechanism to gather *all* messages together to display in a dialog similar to the list of messages shown when loading a profile if anything does not work.