fix: remove parallelization to remove file locking issue
Turns out that running parallel code where an effect of the multiple parts of the code might be reading the same file causes a security exception. It seems that ensuring the profile configuration exists *may* be subject to this, since I can only recreate this problem by running maven, but not ant, or the full application.