feat: allow for different JsonSocketService objects to be returned for different protocol versions
Extends the versioning within the API to allow that a JsonService may use different classes per API version. This allows for the possibliity that a JsonService may have a v5 SocketService, a v6 SocketService, and a common class that these extend from that shares code to reduce complexity in a single class and to reduce duplication between classes.