Update UserInterface.shtml
Page needs to be split into (1) WiFi Throttles page including intro and table and section on remote use, and (2) section on WiThrottle and its UI (the original intent of this page based on the name of the file and where this file resides in the JMRI repository). The changes made here simply alert users to the use of the Web server (rather than WiThrottle) for two of the Wi-Fi throttles listed.