Cross reference use of Raspberry Pi
There is a hardware page related to using Raspberry Pi's GPIO pins.  There is also a separate page for installing JMRI on a Raspberry Pi in the /install directory in the web repository.  I have put in links to cross-reference so that naive user (e.g. me) will realize that these are different and both are possible.