chore: Remove references to classes not in JDK 9 or newer
`com.apple.eawt.AppEvent.*` and `com.apple.eawt.QuitResponse` are not in JDK 9 or newer (even on macOS). This PR removes those classes from the JMRI Java code by placing their use in JavaScripts loaded at compile time.

This dos not remove the need to include AppleJavaExtensions.jar for compilation on non-macOS platforms; use of com.apple.eawt.Application would need to be abstracted out to a script to remove that need.