fix: Add schema blocking use rather than test schema does not exist
`rollingStock` is a JSON type that can be requested to all rolling stock from the Operations database (both cars and engines), but can't be otherwise used. Code/tests/everything turns out to be easier to deal with if a JSON schema for the `rollingStock` client-to-JMRI message that effectively bars use is added than trying to work around not having that schema.