chore: ensure javac creates a class for all package-info files
The current maven (and ant) strategy for determining if recompilation is needed is to compare the list of `.class` files to the list of `.java` sources. Since `package-info.java` does not cause a class to be generated unless it contains an annotation, adding an annotation "tricks" maven into not needlessly recompiling (note that ant has special handling to create a special empty .class file for package-info.java files even if javac would not create a .class file, but maven does not).