feat: Allow a Turnout to follow another Turnout's state
Some features of JMRI expect that two Turnouts will match the known state, not just the commanded state, but when one Turnout is physical with exact feedback, the second Turnout, if internal, will not match the first Turnout's state. This feature allows a Turnout to match another Turnout's state, and is implemented as part of Turnout to avoid the need for other features of JMRI to care if a Turnout is internal or external.