fix: Extend use of VetoableChangeSupport
Make the core subclasses of Manager implement VetoableChangeSupport. (AbstractProxyManager overrides the some methods to ensure the managers that are proxied have the same listeners as it does.)
Deprecate unused functionally duplicative code - PropertyChangeListeners get the same data ManagerDataListeners (in theory) would, and there are PropertyChangeListeners listening for the correct events, and nothing implements ManagerDataListener, so removing it would be a net win for reducing needless code complexity.