LayoutEditor:Remove unnecessary SwingUtilities.invokeLater's
These invokeLater's were unnecessary; the rootPane's could not be found because the buttons parent panels had not yet been added to theContentPane (the root pane). The "fix" was to move the getRootPane/setDefaultButton lines to after the buttons parent panes had been added to theContentPane. Also not-null conditionals were added around the setDefaultButton just in case to avoid NPE's.