chore: prevent instrumentation errors on individual tests
For reasons I do not understand the maven goal `jacoco:prepare-agent` when run as a precursor to the maven target `test` performs dynamic instrumentation, but when run as a specific goal on the maven command line performs static instrumentation.

In any event, there seems to be no appreciable difference in execution speed (given our current set of plugins) between running `test-compile surefire:test` and `test`, so just using `test` to run single-file tests in NetBeans.