Update HornbyTTS.xml
Includes an Update by Marc N Fournier for the Class 7 Britannia that was never submitted but posted on the jmriuser group.  It also included a list of function labels for the Britannia.
Also added is the Merchant Navy and function Labels for all other TTS Diesel and Steam locomotives
Cleaned up any duplicate locos that were hidden from view.