Proposed answer on where to merge Chore PRs
The question was whether "Chore" PRs would go on the master branch.  I think that the master branch should stay clean in this model, so that we can easily do comparisons.  If I've started on master and make a change in bobj-pr, then these are all useful and interesting:
```
git diff master...bobj-pr
git diff bobj-pr...master
git diff bobj-pr...dev-major
git diff bobj-pr...dev-major
git diff bobj-pr...dev-major
```
The first is a check of what I think I'm contributing; the second is a check that I really started from master, and am not bringing in other stuff; and the last three check for conflicts that would prevent merging up the stack of branches.  Those last ones are _much_ easier to interpret if master only moves is a specifically known way.