Fix SprogThrottle 28 step mode EStop
SprogThrottle setSpeedSetting was incorrectly using a value of 0 for EStop when in 28 Step mode.  0 results in a normal stop, while 1 (or 3) results in an EStop.  The comments in the code are correct, just the actual value used was incorrect.  This PR corrects the code to use a value of 1 for EStop in 28 Step mode.

Note: The existing code uses the correct EStop value for 128 Step mode.