Recreate SaveWhatWhereHowWhen.shtml so as to redirect to DataManagement.shtml
This is a kludge to handle the situation where a file was created but its name changed and is still hanging around the web under the old name although no longer in Git.  

This file is being recreated in Git to change its contents to a redirect to the file to which it was renamed.  Once this redirect content gets to the web, the file will again be deleted from Git.