QualifierCombiner fix pane handling
QualifierCombiner propertyChange listener was not getting called when used with Panes (tested on Windows 10). The cause appeared to be that PaneProgPane did not generate propertyChange events when its Enabled state was changed.  In addition the PaneProgPane did generate two propertyChange events on Close which resulted in an unhandled exception in QualifierCombiner.

Changed QualifierCombiner to listen to propertyChange events on the VariableValues used by its component qualifiers.  Also removed the component qualifier propertyChange listeners since those are now redundant.  QualifierCombiner now recalculates properly for both Pane and Variable qualifiers, and does not generate an exception when the template is closed.