Proposed exception to “stale issue” processing
Proposal:  if an issue is explicitly labeled “Bug,” I think it should be exempt from the stale issue processing.  If someone went to the trouble of identifying a verifiable bug, and actually labeling it as such, shouldn’t we at least keep the issue open to give a developer the opportunity to clear it?  Of course, if the label was never applied or is removed, then the issue becomes subject to being closed when stale.

[If there is significant disagreement with this proposal, I will withdraw the PR.]