Initial Commit of SplitEnumVal Dev branch
Implemented SplitEnumVariableValue.java and expanded VariableTableModel to support the new data type.
Currently the enumVal parsing seems functional. Values are correct, drop-downs are in order.
The splitVal components are working in so much as when we trigger a write both CV's are written to.
KNOWN BUGS:
1. Right now we are over-writing the actual retVals (splitting the long into individual bytes) with 128, and 32 respectively. This is just to make assessing the write process easier.
2. At the time of write both CV's are written to but the same value is applied to both.

Expected issues: state management, and overwritten write val