Fix openlcb memory config tool "put" not initializing a the dst variable.
The code in the "pushedPutButton" never initializes the farID variable.
This means that if the user opens the tool and tries to do a put operation,
it never works. The operation would only work if the user does a "get" first,
then followed by a "put".

If the user changes the destination node between a get and put operation,
the put operation will be applied to the destination node that was selected
for the get.