OpenLCB Signal Mast enhancements
1) Support for multiple OpenLCB connections using different prefixes
   than the default "M" prefix. The Add/Edit pane will show a combobox
   selector if there are multiple OpenLCB connections present.
2) Added GUI icons for each signal mast in the Add/Edit signal mast
   pane. This makes it far easier to correlate with the names of the
   signal masts in the signals xml files.

Also changed a few variable names in OlcbSignalMastAddPane.java to be
clearer. This forced a change in the SignalMastAddPaneTest since it uses
the internal object variable names.

It would be good to have a way to test the signal mast add pane with
multiple OpenLCB connections in SignalMastAddPaneTest but I don't see an
easy way to do that. I have tested this manually, both with multiple
OpenLCB simulator connections and with actual real multiple OpenLCB
networks.