Config and Temp not shown same as rest
I noticed there was no space on the line for the Temp file or Config file locations between the colon and the path. Saw only those had the colon in the property but the rest did not. Changed the code and the properties to remove the colon from the properties and make all the code parts the same.