ESU decoder definition update
- Added LokPilot 5 micro DCC Direct, LokPilot 5 nano DCC, LokSound 5 micro DCC Direct Atlas S2 and LokSound 5 nano DCC Next18 (untested, don't have the hardware)
- Added a few productIDs to LokPilot 5 L and LokPilot 5 L DCC
- Moved LokSound 5 micro DCC Direct Atlas Legacy definition to match
  LokProgrammer decoder order
- Removed SUSI function configuration from LokSound 5 micro DCC Direct
  Atlas Legacy

Should fix #13190 from the JMRI decoder definition side. I believe
there's more to rick77337's troubles than just a missing decoder
definition, however...

With this PR, all the LokSound 5s and LokPilot 5s in the
current-as-of-today LokProgrammer 5.2.9 should be detected and work
reasonably well with JMRI with the exception of the LokPilot 5 Basic,
which seems to be a totally different beast from all the others.