Separate finding symlinks and user-specified ports
The prior code found symlinks only on Linux and Mac systems and only
if user-specified pattern was in place, but the pattern was ignored
when a viable symlink was found.  The pattern string was needed to
search for symlinks, but the pattern was not used for matching the
symlink name.

The new code separates the symlink checking and pattern matching.  On
Linux and Mac systems, the files in the /dev directory are searched
first for any viable symlinks, whether or not a pattern is specified,
and then searched for files matching the pattern.  If is either a
symlink or matches the pattern, it is presented as an option for use.