Add EOF detection for LocoNet LnPacketizer
In the prior code, when a USB port was disconnected,
LnPacketizer#readByteProtected would enter a tight loop.  This loop
would never terminate and cause one CPU of the host to be 100%
utilized.

In this change, the end-of-file is detected in
LnPacketizer#readByteProtected, which now throws an EOFException.  The
EOFException is caught in LnPacketizer.RcvHandler#run, which then
terminates the reading and writing threads and disconnects the port.

JMRI still does not automatically recover when the USB port is
reconnected, but it no longer creates a runaway thread and excess load
upon the original disconnect.