Update IdentifyDecoder.java
I am currently constructing an HM7000 TXS decoder definition file and there is a problem with the way JMRI identifies the Product ID - currently class IdentifyDecoder looks at CVs 47,47 and 49 - but these are the firmware definitions and relate to the chip, not the Sound Profile, which is needed to identify the CVs used for each locomotive/sound profile. The CVs that need to be accessed are CV 200 (High Byte) CV 201 (Low Bite) - in most instance the Low Byte (CV201) will be used as the product IDs are all less than 255. I think similar rules could be used as are applied to the TTS chips where CV 158 is an optionalCV.