Update hardware/tmcc/index.shtml
The (Aux1+n) designation is misleading. The Aux1 key only toggles on the numeric keypad using the Lionel remote while in ENG mode. Aux1 is not added onto the code in front of the number (and the Aux1 + n keystroke sequence is needed for some locomotive commands; another element of why this needs to be changed).