ZeroConfService Mgr whitespace hostname
Use the network Identity if the host name is passed as a whitespace String.
This now matches the JavaDoc for the hostName method, i.e. removing leading whitespace.

This now also matches the unit test comment, see amended test line 187 which was previously a duplicate of the test above, i.e. Whitespace to punycode, not just Empty strings to punycode.